Replace API-routes from web.php to api.php
