+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -34,6 +34,7 @@ class AuthenticatedSessionController extends Controller
 
 //        return redirect()->intended(RouteServiceProvider::HOME);
         return redirect('/');
+
     }
 
     /**
@@ -50,6 +51,7 @@ class AuthenticatedSessionController extends Controller
 
         $request->session()->regenerateToken();
 
+//        return response()->json(['status'=>'Log OUT'], 200);
         return redirect('/');
     }
 }
