+++ routes/api.php
@@ -2,6 +2,9 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\TodolistController;
+use App\Http\Controllers\TaskController;
+use App\Http\Controllers\UserController;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,4 +19,44 @@ use Illuminate\Support\Facades\Route;
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
+});
+
+//Lists Routes
+Route::group([
+    'prefix' => '/lists',
+    'middleware' => ['auth']
+], function (){
+    Route::get('/', [TodolistController::class, 'getListsByUser']);
+    Route::get('/{list_id}', [TodolistController::class, 'getListById']);
+    Route::post('/', [TodolistController::class, 'createList'])
+        ->middleware('checkListTitle');
+    Route::put('/{list_id}', [TodolistController::class, 'updateList'])
+        ->middleware('checkListTitle');
+    Route::delete('/{list_id}', [TodolistController::class, 'delete']);
+    Route::get('/prop/predefined', [TodolistController::class, 'getPredefinedLists']);
+});
+
+//Tasks Routes
+Route::group([
+    'prefix' => '/tasks',
+    'middleware' => ['auth']
+], function (){
+    Route::get('/list/{list_id}', [TaskController::class, 'getByListId']);
+    Route::get('/{task_id}', [TaskController::class, 'getById']);
+    Route::post('/', [TaskController::class, 'create'])
+        ->middleware('checkTaskTitle');
+    Route::put('/{task_id}', [TaskController::class, 'update'])
+        ->middleware('checkTaskTitle');
+    Route::delete('/{task_id}', [TaskController::class, 'delete']);
+});
+
+
+//Users Routes
+Route::group([
+    'prefix' => '/api/user',
+//    'middleware' => ['auth']
+], function (){
+    Route::get('/all', [UserController::class, 'getAllUsers']);
+    Route::get('/{user_id}', [UserController::class, 'getUserById']);
+    Route::get('/', 'UserController@getCurrentUser');
 });
